functions/get_expression/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
torchvision==0.14.1+cpu
pillow
opencv-contrib-python-headless
matplotlib
scikit-image
boto3